ARRAY: Interface, Linear search added
Created the interface for array. The interface assumes that array is already
created and does not provide methods for array creation. There is no upper
limit on array length but there is a check for length to be > 0. Interface
does provide functions for insertion and deletion of array elements.

Added implementation of linear search alongside a test for it.