ARRAY: insert, delete added
Implementation and interface added for insert and delete of an
array element.

Both insert and delete function do not change the length of the
array. While inserting, the last element of the array is
discarded and while deleting, 0 is placed insted of the element
to be deleted. The elements are swapped and placed in their
right position ofcourse.