LINKLIST: new, insert, print and count added
Implementations and intefrace for creation, insertion and
printing of link list added.

In the insert function:
int LinkList_insert(LinkList *list, int item, LinkNode *node);

It inserts a new node after the node specified. If node is
null, then it will insert to the end of the list. Should also
have a function which will insert the node before the
specified node and to the begining of list if node == null.