BSTree: Interface added
Interface for binary search tree added. Just like the linked
list, it is not an ADT because it is meant to be used as
a building block for other ADTs with full control of its
nodes. The only real problem was traversing the nodes as per
requirements, which won't be possible if it's an ADT.