HEAP: INCOMPLETE: Problem Implementing
Marked heap as incomplete because of problem implementing it.
The main problem is to figure out how to do insertion of a
new element in heap. Algorithm is understood, but converting it
to code is a problem. Unable to figure out how to find out the
location of the new element to be inserted, before it is to be
swaped to it's correct place.